Fix -Wabsolute-value warning

+++ src/DETHRACE/common/racesumm.c
@@ -957,7 +957,7 @@ int DamageScrnUp(int* pCurrent_choice, int* pCurrent_mode) {
                         new_selection = i;
                         break;
                     }
-                    difference = abs(gWreck_array[i].pos_x - gWreck_array[gWreck_selected].pos_x);
+                    difference = abs((int)(gWreck_array[i].pos_x - gWreck_array[gWreck_selected].pos_x));
                     if (difference < new_difference) {
                         new_selection = i;
                         new_difference = difference;
